[CDF-28413] Move deployment dependency detection to CRUDs
